feat: allow multiple attempts to fail before caching a negative result (#140)
Requests for feature flags get cancelled if there's a page change around the same time as the flag request goes out. There's no reliable way to know if a request was cancelled or otherwise failed.

This PR changes the caching logic so we will not use a negative cached result until it's failed multiple (configurable) times. If any of those attempts succeed, that response will be used. Finally, if they all fail, a negative response will be cached for a short while to avoid having the user wait for more failures and spamming the API.