Node.js SDK batching support (#213)
This PR adds a new class into the mix. This class is used as a "batch
buffer" that will store all given data and then flush it at a given time
interval or when the buffer is full.

There is also a secondary retry buffer that is being flushed on a
different cadence and items in there are dropped out after a number of
tries.

The `BucketClient` has been updated to route `user`, `company` and
`feature-flag-event` payloads into the buffer instead of calling the
individual API endpoints.