fix: set staletime to zero to avoid confusion (#232)
At initialization time (page load) we fetch features. There's a cache of
60s today which means if you loaded the page previously within 60s we
will not refetch. This can be cause for confusion. After changing
targeting rules, people expect that after a hard refresh features will
have updated. If they wait a bit and then hard refresh they do change
which makes them end up looking random. We added this cache to avoid
help in the case of non-SPA which would fetch features with every load
and incur the overhead of the fetch request every time. However, since
the norm is an SPA I'm changing it to default to zero. Folks with an
non-SPA can still set the stale time to 60s and avoid the refetch with
every load.