feat(node-sdk): Implement automatic attribute sending and tracking toggle (#235)
This PR adds automatic sending of `user` and `company` events to Bucket
when either `bindClient` or `getFeature(s)` calls are made.

Also introduces `enableTracking` in `Context` to allow disabling to send
`event`, `user` and `company` when the user is not interested in sending
them (e.g. impersonating other users).

Employing the rate-limiter on `user` and `company` calls to limit the
amount of potential calls made to `Bucket` servers.