Toolbar (alpha) (#290)
The toolbar enables teams to quickly enable/disable features directly in
their own app.

It's a UI that renders on top of the page:
<img width="265" alt="image"
src="https://github.com/user-attachments/assets/18e0b483-68b8-43cb-996c-2a80b9aa6fe4"
/>

Features so far:
- fetches features remotely through the SDK
- also takes a list of features passed locally by prop
- override feature enabled easily by flipping a switch, which is stored
in localstorage
- filter the list of features by search term
- works for both Browser SDK and React SDK

This started out as a way to see just how hard it would be to implement
a toolbar for Bucket, so the code is now really in great shape.

Known issues:
- we use `featureList` to pass the string[] with feature keys into the
BuckerProvider (and from there to BucketClient from the browser-sdk
because `features` was already take for misc. feature fetching related
options
- positioning (bottom-left etc.) of the toolbar toggle doesn't really
work the the toolbar dialog open up weirdly/wonkily in lots of situation
(like narrow screen).

---------

Co-authored-by: Lasse Boisen Andersen <la@laander.com>
Co-authored-by: Lasse Boisen Andersen <lasse@bucket.co>