Upgrade ESLint and Prettier (#336)
To use ESM modules properly in the CLI package ESLint needed an upgrade.
This also uncovered a handful of omissions in the code. I have divided
up the commits to help facilitate review:

1. The first commit is pure style changes due to the React plugin
https://github.com/bucketco/bucket-javascript-sdk/commit/e42cce56e2361cdc13f999d9f666f8d6ad0d8f5f.
2. The second commit has the bulk of the configuration and some
important fixes
https://github.com/bucketco/bucket-javascript-sdk/commit/9c05eb5ca82974602acc7fd57b32fceb288abbe0.
3. The third commit has some business logic changes due to the incorrect
circular reference
https://github.com/bucketco/bucket-javascript-sdk/commit/0f9df07674b9363913bc9e184df0491cf2a2dce9.
4. The fourth commit is to handle the preact libraries
https://github.com/bucketco/bucket-javascript-sdk/pull/336/commits/5af8eb75622b59939c6b25ac04eb5f660778dbbc

The second and third commits are the important to review and I would
love help testing the feedback modal to make sure no infinite loops have
crept in as is so common with `useEffect` and `function` dependencies.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Matus Vacula <matus@bucket.co>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Alexandru Ciobanu <alex+git@ciobanu.org>
Co-authored-by: Rasmus Makwarth <makwarth@gmail.com>
Co-authored-by: Ron Cohen <ron@bucket.co>