Toolbar: handle many features, tweak stagger anim (#349)
Handle many features:


https://github.com/user-attachments/assets/0f828aca-9544-497a-acdb-02fff4761668

We show a staggered entrance animation for each feature. But if the
stagger delay is static and you have a lot of features, that might
result in a (too) long animation. So instead, we adjust the stagger
delay according to the amount of features in the list so the total
animation duration never exceeds 75ms. Here's a demonstration:


https://github.com/user-attachments/assets/ea3c110c-d8e6-443f-b132-8b174d734734