feat: use new login scheme which supports Safari (#375)
This authentication flow does not use a fetch request to
`http://localhost` but instead redirects there. We don't want to pass
the actual access token along so it becomes part of the browser history.
Instead we supple a code which is then exchanged for the actual access
token through a new API endpoint in the backend.

This starts to look very much like a standard oauth flow. The key
missing piece is come `client_id` parameters. I'm making the new server
endpoints live under `/api/oauth/cli` so that when we do introduce
propert oauth there will be no clashes.

I removed the spinner on login because it indicates that it's working
when in reality it's waiting for you to do something. I'm also printing
out the URL in case opening of the browser doesn't work.