fix(node-sdk): increase default `get` timeout to 10s, add 3 retries (#388)
Our HTTP API responds to `/feature` requests in less than 5ms, but there
are cases where the network on the client could be dodgy or there's
otherwise some delay. Especially when the program is just starting up
this is problematic because the NodeSDK will use the fallback flags
which for the most part are empty.

This introduces a configurable timeout and default the timeout to 10s
instead of the existing 5s. It also adds an exponential retry which is
set to 3 tries by default.