fix: Improve logging when not initialized (#410)
Correctly initializing the BucketClient can be a bit tricky. This should
complain more loudly when it happens.

I considered and rejected the idea of throwing an error when this
happens because incorrectly constructed initialization code may only
break under high concurrency and thus only show up in production which
could cause down time if we throw.